separated state management code
